feat: useScrolledPastThreshold 커스텀 훅
* 스크롤이 thresholdY 값을 지났는지 확인하는 hook

* Header 컴포넌트에 있던 로직 분리
